+++ apps/E2E/src/LinkExperimental/pages/LinkPageObject.ts
@@ -3,7 +3,7 @@ import {
   EXPERIMENTAL_LINK_TEST_COMPONENT,
   HOMEPAGE_EXPERIMENTAL_LINK_BUTTON,
   EXPERIMENTAL_LINK_NO_A11Y_LABEL_COMPONENT,
-} from '../../../../fluent-tester/src/TestComponents/LinkExperimental/consts';
+} from '../consts';
 import { BasePage, By } from '../../common/BasePage';
 
 class ExperimentalLinkPageObject extends BasePage {