+++ apps/fluent-tester/src/TestComponents/Menu/MenuTest.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import { ButtonV1 as Button } from '@fluentui/react-native';
 import {
   Menu,
@@ -12,13 +12,14 @@ import {
 } from '@fluentui-react-native/menu';
 import { Stack } from '@fluentui-react-native/stack';
 import { stackStyle } from '../Common/styles';
-import { MENU_TESTPAGE } from './consts';
+import { MENU_TESTPAGE } from '../../../../E2E/src/Menu/consts';
 import { Test, TestSection, PlatformStatus } from '../Test';
 import { TextV1 as Text } from '@fluentui-react-native/text';
 import { E2EMenuTest } from './E2EMenuTest';
-import { StyleSheet } from 'react-native';
 import { MenuTriggerHoverCallback, MenuTriggerOnClickCallback } from './MenuTriggerCallbacks';
 import { MenuTriggerChildRef } from './MenuRefs';
+import { StyleSheet } from 'react-native';
+import { MenuScrollView } from './MenuScrollView';
 
 const MenuDefault: React.FunctionComponent = () => {
   return (
@@ -268,6 +269,10 @@ const menuSections: TestSection[] = [
   {
     name: 'Menu Submenu',
     component: MenuSubMenu,
+  },
+  {
+    name: 'Menu with ScrollView',
+    component: MenuScrollView,
   },
   {
     name: 'Menu Trigger onClick Override',