+++ apps/fluent-tester/src/TestComponents/Tokens/index.ts
@@ -1,2 +1,1 @@
 export * from './TokenTest';
-export * from './consts';