+++ packages/components/ContextualMenu/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log - @fluentui-react-native/contextual-menu
 
-This log was last generated on Thu, 15 Dec 2022 14:50:50 GMT and should not be manually modified.
+This log was last generated on Tue, 27 Dec 2022 22:21:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.21.7
+
+Tue, 27 Dec 2022 22:21:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.21.45
+- Bump @fluentui-react-native/focus-zone to v0.11.15
+- Bump @fluentui-react-native/icon to v0.16.5
+- Bump @fluentui-react-native/interactive-hooks to v0.22.3
+- Bump @fluentui-react-native/text to v0.19.2
+- Bump @fluentui-react-native/tokens to v0.20.1
+- Bump @uifabricshared/foundation-compose to v1.12.29
+- Bump @fluentui-react-native/pressable to v0.9.42
+
+## 0.21.6
+
+Tue, 27 Dec 2022 10:18:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.21.44
+- Bump @fluentui-react-native/focus-zone to v0.11.14
+- Bump @fluentui-react-native/icon to v0.16.4
+- Bump @fluentui-react-native/interactive-hooks to v0.22.2
+- Bump @fluentui-react-native/text to v0.19.1
+- Bump @uifabricshared/foundation-compose to v1.12.28
+- Bump @fluentui-react-native/pressable to v0.9.41
+
+## 0.21.5
+
+Thu, 22 Dec 2022 19:31:49 GMT
+
+### Patches
+
+- Fork SubmenuItem pressable (amchiu@microsoft.com)
 
 ## 0.21.4