+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,37 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Thu, 15 Dec 2022 14:50:50 GMT and should not be manually modified.
+This log was last generated on Tue, 27 Dec 2022 22:21:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.7
+
+Tue, 27 Dec 2022 22:21:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.7
+- Bump @fluentui-react-native/experimental-button to v0.16.68
+- Bump @fluentui-react-native/framework to v0.8.28
+- Bump @fluentui-react-native/tokens to v0.20.1
+
+## 0.6.6
+
+Tue, 27 Dec 2022 10:18:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.6
+- Bump @fluentui-react-native/experimental-button to v0.16.67
+- Bump @fluentui-react-native/framework to v0.8.27
+
+## 0.6.5
+
+Thu, 22 Dec 2022 19:31:49 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.5
 
 ## 0.6.4